# =act improve ByteString#take(...)