### 
Currently, we use `ByteStringBuilder` to create a new `ByteString` instance, which would not be quite efficient.

Instead of doing this, we can do as follows so that we can achieve better performance:
1. Seek the index of _last_ vector element we need to _take_
2. Find the number of characters left to take from the _last_ `ByteString1` element.
3. Create `ByteString` based on the information we obtained from 1 and 2

Then we just need to create a new `Vector[ByteString1]` at most twice, which should be better than the current implementation, i.e., _append_ a new element every time we check `bytestrings(Vector[ByteString1])` element, which ends up O(N) _append_ execution where _N_ is the length of `bytestrings`.

### Benchmark
I measured the performance using sbt-jmh with this [code](https://gist.github.com/monkey-mas/1ec7a4fca2a54511b4c8366f80adf94c).  Benchmark code was executed with some parameters such as `-t1 -f2 -wi 10 -i 20`.

#### Test environment
```
OS: Ubuntu 14.0.4 (Linux kernel version 3.13.0-95-generic)
CPU: Intel(R) Core(TM) i7-2600 CPU @ 3.40GHz
VM version: JDK 1.8.0_77
```

#### Before
```
[info] Take.take_avg                   thrpt   40       28804.052 ±      567.324  ops/s
[info] Take.take_best                  thrpt   40    21372775.581 ±   726227.926  ops/s
[info] Take.take_greater_or_eq_to_len  thrpt   40   896854941.095 ±  5621572.730  ops/s
[info] Take.take_negative              thrpt   40  1173291248.506 ± 13086338.839  ops/s
[info] Take.take_worst                 thrpt   40       12358.863 ±      535.931  ops/s
```

#### After
```
[info] Take.take_avg                   thrpt   40      567539.273 ±    2700.264  ops/s
[info] Take.take_best                  thrpt   40   207257188.976 ± 1369415.345  ops/s
[info] Take.take_greater_or_eq_to_len  thrpt   40   929926474.479 ± 5372591.197  ops/s
[info] Take.take_negative              thrpt   40  1244393296.012 ± 6830415.239  ops/s
[info] Take.take_worst                 thrpt   40      252899.243 ±    6644.853  ops/s

```