@johanandren 2016-09-12
I think it would be nice with the benchmarks included in the `akka-bench-jmh` module of Akka though.