@ktoso 2016-09-12
Please include the benchmark code in the existing benchmarks too (in akka-benchm-jmh).
Some minor cleanups and LGTM then, thanks for picking this up!