@monkey-mas 2016-09-12
Thanks for quick your reviews!

> Please include the benchmark code in the existing benchmarks too (in akka-benchm-jmh)

Okay I'll add my benchmark by this weekend for sure!! :)