@ktoso 2016-09-22
Thanks for the meet work on those!
Feel free to.include benchmarks in such PRs as well btw, always useful to extend the suite and understand the results pasted.