@monkey-mas 2016-09-22
Oops sorry for not finishing my homework before you merge it. I've been busy day in and day out recently :(
Anyway, thanks for your help! :) will send PRs! :dancers: